requirements.txt
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
faiss==1.8.0
fastapi==0.115.12
langchain==0.3.19
langchain-groq==0.3.2
langchain-huggingface==0.1.0
langgraph==0.4.2
langsmith==0.1.147
pydantic==2.9.2
torch==2.4.1
torchaudio== 2.7.0
torchvision==0.19.1
transformers==4.45.1
unstructured==0.16.5
jinja2==3.1.5
huggingface_hub==0.30.2
llama-cloud==0.1.13
llama-cloud-services==0.6.3
nest-asyncio==1.6.0